requirements.txt
## Minimal pinned dependencies to satisfy the Constitution
pymodbus==3.1.0
influxdb-client==1.33.0
PyYAML==6.0
python-dotenv==1.0.0
pytest>=7.0

